getting timezone for better forecast

+++ models/weather.go
@@ -25,14 +25,14 @@ type Hourly struct {
 	Time                []string  `json:"time"`
 	Temperature         []float64 `json:"temperature_2m"`
 	FeelsLike           []float64 `json:"apparent_temperature"`
-	PrecipitationChance []int     `json:"precipitation_probability"`
-	Rain                []int     `json:"rain"`
-	Showers             []int     `json:"showers"`
-	Snowfall            []int     `json:"snowfall"`
-	SnowDepth           []int     `json:"snow_depth"`
-	CloudCover          []int     `json:"cloud_cover"`
+	PrecipitationChance []float64 `json:"precipitation_probability"`
+	Rain                []float64 `json:"rain"`
+	Showers             []float64 `json:"showers"`
+	Snowfall            []float64 `json:"snowfall"`
+	SnowDepth           []float64 `json:"snow_depth"`
+	CloudCover          []float64 `json:"cloud_cover"`
 	WindSpeed           []float64 `json:"wind_speed_10m"`
-	WindDirection       []int     `json:"wind_direction_10m"`
+	WindDirection       []float64 `json:"wind_direction_10m"`
 	SoilTemperature     []float64 `json:"soil_temperature_0cm"`
 }
